--- services/passenger-service/src/main/java/com/Harevich/passenger_service/PassengerServiceApplication.java
@@ -1,13 +1,0 @@
-package com.Harevich.passenger_service;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class PassengerServiceApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(PassengerServiceApplication.class, args);
-	}
-
-}
