--- services/passenger-service/src/main/java/com/Harevich/passenger_service/util/constants/RegularConstants.java
@@ -1,8 +1,0 @@
-package com.Harevich.passenger_service.util.constants;
-
-import lombok.experimental.UtilityClass;
-
-@UtilityClass
-public class RegularConstants {
-    public final String PHONE_NUMBER_REGEX = "\\+375(29|44|25|33|17)\\d{7}";
-}
