+++ services/passenger-service/src/main/java/com/Harevich/passengerservice/PassengerServiceApplication.java
@@ -1,0 +1,13 @@
+package com.Harevich.passengerservice;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class PassengerServiceApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(PassengerServiceApplication.class, args);
+	}
+
+}
