+++ services/passenger-service/src/main/java/com/Harevich/passengerservice/exceptions/UniqueException.java
@@ -1,0 +1,10 @@
+package com.Harevich.passengerservice.exceptions;
+
+import lombok.EqualsAndHashCode;
+import lombok.RequiredArgsConstructor;
+
+@EqualsAndHashCode(callSuper = true)
+@RequiredArgsConstructor
+public class UniqueException extends RuntimeException{
+    private final String message;
+}
