+++ services/passenger-service/src/main/java/com/Harevich/passengerservice/model/Passenger.java
@@ -1,0 +1,38 @@
+package com.Harevich.passengerservice.model;
+
+import jakarta.persistence.*;
+import lombok.*;
+import org.hibernate.annotations.CreationTimestamp;
+import org.hibernate.annotations.UpdateTimestamp;
+
+import java.time.LocalDateTime;
+import java.util.UUID;
+
+@Entity
+@AllArgsConstructor
+@NoArgsConstructor
+@Getter
+@Builder
+@Setter
+@Table(name = "passenger")
+public class Passenger {
+    @Id
+    @GeneratedValue(strategy = GenerationType.UUID)
+    private UUID id;
+    @Column(nullable = false)
+    private String name;
+    @Column(nullable = false)
+    private String surname;
+    @Column(unique = true,nullable = false)
+    private String email;
+    @Column(unique = true,nullable = false)
+    private String number;
+    @CreationTimestamp
+    @Column(name = "created_at", updatable = false)
+    private LocalDateTime createdAt;
+    @UpdateTimestamp
+    @Column(name = "updated_at")
+    private LocalDateTime updatedAt;
+
+    private boolean deleted = false;
+}
