+++ services/passenger-service/src/main/java/com/Harevich/passengerservice/util/additional/PassengerValidation.java
@@ -1,0 +1,5 @@
+package com.Harevich.passengerservice.util.additional;
+
+public interface PassengerValidation {
+    public void handleUniqueExceptionByErrorMessage(String rootMessage);
+}
