+++ services/passenger-service/src/main/java/com/Harevich/passengerservice/util/constants/PassengerServiceResponseConstants.java
@@ -1,0 +1,12 @@
+package com.Harevich.passengerservice.util.constants;
+
+import lombok.experimental.UtilityClass;
+
+@UtilityClass
+public class PassengerServiceResponseConstants {
+    public final String PASSENGER_NOT_FOUND = "Passenger with such id not found";
+    public final String PASSENGER_DELETED = "Passenger with such id is deleted";
+
+    public final String REPEATED_EMAIL = "Passenger with such email already exists";
+    public final String REPEATED_PHONE_NUMBER = "Passenger with such phone number already exists";
+}
