+++ services/passenger-service/src/main/java/com/Harevich/passengerservice/util/constants/RegularConstants.java
@@ -1,0 +1,8 @@
+package com.Harevich.passengerservice.util.constants;
+
+import lombok.experimental.UtilityClass;
+
+@UtilityClass
+public class RegularConstants {
+    public final String PHONE_NUMBER_REGEX = "\\+375(29|44|25|33|17)\\d{7}";
+}
